Organize Initializer.kt
Inline InteractionHandler.kt
